Update the intro page and updated CV
Eliminated the previous content of the intro page and added a paragraph describing me and my work. Also added a link to the website in my CV. Finally, I added an invisible line at the top of the intro page with my name so that this is the first thing that appears in the tab title of the intro page.